add: more e2e tests
