+++ pruebas/01-reading-list/aitorjs/src/App.tsx
@@ -79,7 +79,7 @@ function App() {
           ))}
         </div>
 
-        <div className="flex flex-col">
+        <div className="flex flex-col" id="readList">
           {books.length}
           <ReadList />
         </div>
